+++ server.js
@@ -3,7 +3,7 @@ const jwt = require('jwt-simple');
 const express = require("express");
 const app = express();
 
-app.use(express.static(__dirname + '/build'));
+app.use(express.static(__dirname + '/build'));''
 app.use(require('body-parser').json({limit: '25mb'}));
 app.use(require('cookie-parser')());
 
@@ -34,7 +34,7 @@ app.use((req, res, next) => {
 
 
 app.use(require('./server/homebrew.api.js'));
-//app.use(require('./server/admin.api.js'));
+app.use(require('./server/admin.api.js'));
 
 
 const HomebrewModel = require('./server/homebrew.model.js').model;